+++ carbonalysis-api/src/main/java/com/github/carbonalysis/domains/emissions/EmissionsRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface EmissionsRepository extends JpaRepository<EmissionsData, Long> {
-  List<EmissionsData> findByUserId(Long id);
+  List<EmissionsData> findByUserId(int id);
 }
